# docs/openapi/catalogo.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Catálogo de filmes — API pública",
    "version": "1.0.0",
    "description": "Sessão pelo cookie HttpOnly token. Faça login em /login.html no mesmo navegador ou execute POST /api/auth/login com uma conta local. Não cole o cookie no Authorize. GET /api/auth/me sem sessão retorna 401 e é uma chamada segura para demonstrar Try it out. Papel comum: usuario; administrador: admin. Alterações de dados devem ser testadas somente no ambiente local."
  },
  "servers": [
    {
      "url": "/",
      "description": "Mesma origem do catálogo (local ou publicada)"
    }
  ],
  "components": {
    "schemas": {
      "Mensagem": {
        "type": "object",
        "properties": {
          "mensagem": {
            "type": "string"
          }
        },
        "required": [
          "mensagem"
        ]
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "Pessoa de exemplo"
          },
          "email": {
            "type": "string",
            "example": "pessoa@example.invalid"
          },
          "role": {
            "type": "string",
            "enum": [
              "usuario",
              "admin"
            ],
            "example": "usuario"
          }
        },
        "required": [
          "id",
          "nome",
          "email",
          "role"
        ]
      },
...
      "Role": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "usuario",
              "admin"
            ],
            "example": "usuario"
          }
        },
        "required": [
          "role"
        ]
      },
...
    },
    "securitySchemes": {
      "sessao": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "Cookie HttpOnly criado por POST /api/auth/login. Faça login pelo formulário /login.html no mesmo navegador ou execute o login no Swagger. O navegador envia o cookie automaticamente; não leia nem cole o cookie no Authorize. SameSite=Lax; Secure em produção; duração 8 horas."
      }
    }
  },
  "paths": {
...
    "/api/admin/usuarios": {
      "get": {
        "tags": [
          "Administração"
        ],
        "summary": "Listar usuários (admin)",
        "description": "Exige role admin; retorna somente id, nome, email e role, em ordem de nome e id. Requer sessão autenticada (cookie HttpOnly token).",
        "operationId": "get_api_admin_usuarios",
        "security": [
          {
            "sessao": []
          }
        ],
        "responses": {
          "200": {
            "description": "Usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usuarios": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Usuario"
                      }
                    }
                  },
                  "required": [
                    "usuarios"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Usuário não autenticado. / Token não informado. / Sessão inválida ou expirada. / Token com usuário inválido. / Usuário não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Usuário não autenticado."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Token não informado."
                    }
                  },
                  "2": {
                    "value": {
                      "mensagem": "Sessão inválida ou expirada."
                    }
                  },
                  "3": {
                    "value": {
                      "mensagem": "Token com usuário inválido."
                    }
                  },
                  "4": {
                    "value": {
                      "mensagem": "Usuário não encontrado."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso permitido somente a administradores.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Acesso permitido somente a administradores."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Erro interno do servidor."
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Serviço de autenticação indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Serviço de autenticação indisponível."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/usuarios/{id}/role": {
      "patch": {
        "tags": [
          "Administração"
        ],
        "summary": "Alterar papel (admin)",
        "description": "Exige admin atual no banco; ID em dígitos decimais positivos e Number.isSafeInteger. Não permite alterar o próprio papel nem remover o último admin. Mensagem: Papel do usuário atualizado com sucesso. ou Usuário já possui o papel solicitado. Requer sessão autenticada (cookie HttpOnly token).",
        "operationId": "patch_api_admin_usuarios_id_role",
        "security": [
          {
            "sessao": []
          }
        ],
        "responses": {
          "200": {
            "description": "Papel alterado ou já atribuído",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensagem": {
                      "type": "string"
                    },
                    "usuario": {
                      "$ref": "#/components/schemas/Usuario"
                    }
                  },
                  "required": [
                    "mensagem",
                    "usuario"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "ID do usuário inválido. / Role deve ser usuario ou admin. JSON malformado ou corpo inválido para express.json (resposta padrão HTML do Express).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "ID do usuário inválido."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Role deve ser usuario ou admin."
                    }
                  }
                }
              },
              "text/html": {
                "schema": {
                  "type": "string",
                  "description": "Página padrão de erro do Express; corpo depende do ambiente."
                }
              }
            }
          },
          "401": {
            "description": "Usuário não autenticado. / Token não informado. / Sessão inválida ou expirada. / Token com usuário inválido. / Usuário não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Usuário não autenticado."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Token não informado."
                    }
                  },
                  "2": {
                    "value": {
                      "mensagem": "Sessão inválida ou expirada."
                    }
                  },
                  "3": {
                    "value": {
                      "mensagem": "Token com usuário inválido."
                    }
                  },
                  "4": {
                    "value": {
                      "mensagem": "Usuário não encontrado."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso permitido somente a administradores.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Acesso permitido somente a administradores."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Usuário não encontrado."
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Não é possível remover o papel do último administrador do sistema. / Não é permitido alterar o próprio papel administrativo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Não é possível remover o papel do último administrador do sistema."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Não é permitido alterar o próprio papel administrativo."
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "Corpo acima do limite padrão de 100 KiB de express.json/urlencoded (HTML do Express).",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "description": "Página padrão de erro do Express; corpo depende do ambiente."
                }
              }
            }
          },
          "415": {
            "description": "Charset ou Content-Encoding não suportado pelo parser (HTML do Express).",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "description": "Página padrão de erro do Express; corpo depende do ambiente."
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Erro interno do servidor."
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Serviço de autenticação indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Serviço de autenticação indisponível."
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Dígitos decimais sem zeros à esquerda; inteiro positivo seguro JavaScript.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Role"
              }
            }
          }
        }
      }
    }
  },
...
}